Working on area daily cell
